dom/events/test/test_wheel_zoom_on_form_controls.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Zoom using wheel should work on form controls</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<button id="button" style="width:10px;height:10px;"></button><br>
<input id="input" style="border: 10px solid green;"><br>
<textarea id="textarea" style="border: 10px solid green;"></textarea><br>
<select id="select"><option></option></select>
<script>

  function testControl(id) {
    return new Promise(function(resolve) {
      var initialZoom = SpecialPowers.getFullZoom(window);
      var element = document.getElementById(id);
      element.onwheel = function() {
        window.requestAnimationFrame(function() {
          setTimeout(function() {
            isnot(SpecialPowers.getFullZoom(window), initialZoom,
                  "Should have zoomed");
            SpecialPowers.setFullZoom(window, initialZoom);
            setTimeout(resolve);
          });
        });
      }

      let event = {
        deltaMode: WheelEvent.DOM_DELTA_LINE,
        deltaY: 3,
        ctrlKey: true
      };
      synthesizeWheel(element, 5, 5, event);
    });
  }

  async function test() {
    await testControl("button");
    await testControl("input");
    await testControl("textarea");
    await testControl("select");
    SimpleTest.finish();
  }

  SimpleTest.waitForExplicitFinish();
  SimpleTest.waitForFocus(test);
</script>
</body>
</html>
